cli: rename synctest to syncbench
That better reflects what it's doing and I'm adding an actual synctest
utility.

Release note: None
