+++ pkg/cli/syncbench/syncbench.go
@@ -13,7 +13,7 @@
 // permissions and limitations under the License. See the AUTHORS file
 // for names of contributors.
 
-package synctest
+package syncbench
 
 import (
 	"context"
